Added focus blur plug-in

+++ build-appimage-plugins.sh
@@ -16,6 +16,17 @@ export PATH=${GIMPPREFIX}/bin:$PATH
 # tar xf gutenprint-5.3.1.tar.xz && cd gutenprint-5.3.1 && ./configure CFLAGS='-g -O2' CXXFLAGS='-g -O2' --prefix=/zyx && make install) || exit 1
 # exit
 
+if [ x"${GTK_VERSION}" != "x3" ]; then
+
+gimplibdir=$(pkg-config --variable=gimplibdir gimp-2.0)
+echo "gimplibdir: $gimplibdir"
+if [ -z "$gimplibdir" ]; then exit 1; fi
+
+(cd /work && rm -rf gimp-focusblur-plugin && git clone https://github.com/JMoerman/gimp-focusblur-plugin.git && \
+mkdir gimp-focusblur-plugin/build && cd gimp-focusblur-plugin/build && \
+cmake -DCMAKE_BUILD_TYPE=Release -DPLUGIN_DIR="$gimplibdir/plug-ins" .. && \
+make -j 2 install) || exit 1
+
 
 yum install -y qt5-qtbase-devel qt5-linguist libcurl-devel || exit 1
 (cd /work && rm -rf gmic gmic-qt && \
@@ -24,26 +35,8 @@ git clone https://github.com/dtschump/gmic.git gmic-clone && \
 make -C gmic-clone/src CImg.h gmic_stdlib.h && \
 qmake-qt5 QMAKE_CFLAGS+="${CFLAGS} -O2" QMAKE_CXXFLAGS+="${CXXFLAGS} -O2" CONFIG+=Release HOST=gimp GMIC_PATH=gmic-clone/src && \
 make -j 3 && make install) || exit 1
-
-#cmake .. -DGMIC_QT_HOST=gimp -DCMAKE_BUILD_TYPE=Release && make) || exit 1
-
-gimplibdir=$(pkg-config --variable=gimplibdir gimp-2.0)
-echo "gimplibdir: $gimplibdir"
-if [ -z "$gimplibdir" ]; then exit 1; fi
-
 mkdir -p "$gimplibdir/plug-ins" || exit 1
 cp -a /work/gmic-qt/gmic_gimp_qt "$gimplibdir/plug-ins" || exit 1
-
-
-yum install -y gnome-common
-(cd /work && rm -rf gtkimageview && git clone https://github.com/aferrero2707/gtkimageview.git && \
-cd gtkimageview && patch -N -p0 < /sources/gtkimageview-Werror.patch && \
-./autogen.sh --prefix=/zyx && make -j 2 install) || exit 1
-(cd /work && rm -rf nufraw* && \
-wget https://launchpad.net/~dhor/+archive/ubuntu/myway/+sourcefiles/nufraw/0.42-1dhor~xenial/nufraw_0.42.orig.tar.xz && \
-tar xvf nufraw_0.42.orig.tar.xz && cd nufraw-0.42 && ./autogen.sh && \
-./configure --enable-contrast --prefix=/usr/local && make -j 2 install) || exit 1
-
 
 (cd /work && rm -rf liblqr && git clone https://github.com/carlobaldassi/liblqr.git && \
 cd liblqr && ./configure --prefix=/zyx && make -j 2 install) || exit 1
@@ -57,13 +50,30 @@ cmake -DBABL_FLIPS_DISABLED=OFF -DCMAKE_BUILD_TYPE=Release .. && \
 make -j 2 install) || exit 1
 
 
+
+yum install -y gnome-common
+(cd /work && rm -rf gtkimageview && git clone https://github.com/aferrero2707/gtkimageview.git && \
+cd gtkimageview && patch -N -p0 < /sources/gtkimageview-Werror.patch && \
+./autogen.sh --prefix=/zyx && make -j 2 install) || exit 1
+(cd /work && rm -rf nufraw* && \
+wget https://launchpad.net/~dhor/+archive/ubuntu/myway/+sourcefiles/nufraw/0.42-1dhor~xenial/nufraw_0.42.orig.tar.xz && \
+tar xvf nufraw_0.42.orig.tar.xz && cd nufraw-0.42 && ./autogen.sh && \
+./configure --enable-contrast --prefix=/usr/local && make -j 2 install) || exit 1
+
+
 (cd /work && rm -rf resynthesizer && git clone https://github.com/bootchk/resynthesizer.git && \
 cd resynthesizer && ./autogen.sh --prefix=/zyx && make -j 2 install) || exit 1
+
+fi
 
 
 
 echo "Installing the GIMP scripts"
-gimpdatadir=$(pkg-config --variable=gimpdatadir gimp-2.0)
+if [ x"${GTK_VERSION}" != "x3" ]; then
+    gimpdatadir=$(pkg-config --variable=gimpdatadir gimp-3.0)
+else
+    gimpdatadir=$(pkg-config --variable=gimpdatadir gimp-2.0)
+fi
 (cd /work && rm -rf gimp-scripts && mkdir -p gimp-scripts && cd gimp-scripts && \
 wget https://github.com/pixlsus/GIMP-Scripts/archive/master.zip && \
 unzip master.zip && mkdir -p "$gimpdatadir/scripts" && \
